Got userdata started and sorta working. Wanna se if it is ok to get rid of the lazy static stuff though

+++ shooter-client/src/fps_counter.rs
@@ -1,6 +1,6 @@
 pub struct FpsCounter {
     frame_count: i32,
-    accum_time: f64,
+    accum_time: f32,
 }
 
 impl FpsCounter {
@@ -11,7 +11,7 @@ impl FpsCounter {
         }
     }
 
-    pub fn update(&mut self, dt: f64) {
+    pub fn update(&mut self, dt: f32) {
         self.frame_count += 1;
         self.accum_time += dt;
